feat(create_raw.py): Added a function to Creator to handle yamcs parameters and refactored DB handling.
Was previously too atomic by instantiating an engine for every DB
interation, now an engine and a sessionmaker are expected at the top of
scope, and sessions are passed into functions.  Created a from_yamcs_parameters() function that has a hard-coded set of key/value pairs that will hopefully decrease over time as we learn more things.  Improved timestamp and bit depth handling in tif_info() and removed _template() as it wasn't really needed.